refactor(resources): unify disk resource into single cross-platform file
Merge disk_darwin.go and disk_linux.go into disk.go since both
implementations are nearly identical — the only difference was
syscall.Statfs (linux) vs unix.Statfs (darwin). Use unix.Statfs
from golang.org/x/sys/unix which works on both platforms.

Addresses review feedback from @sjmiller.
